Ported @jbpayne's dashboard code
Moved the code that Jonathan put in navigation/MainTabNavigator.js
into screens/dashboard/index.js as @johnneed suggested.

I commented out the existing global style import since it conflicted
with Jonathan's styles and I was trying to make as few chages as possible.
I left a todo there though.

The app is currently running fine and displaying the new dashboard content.
